# docs/swagger/swagger.yaml
basePath: /api/v1
definitions:
  dto.AddItemRequest:
    properties:
      product_id:
        type: string
      quantity:
        type: integer
    required:
    - product_id
    - quantity
    type: object
...
  dto.GenerateRequest:
    properties:
      code:
        maxLength: 50
        minLength: 3
        type: string
      discount_type:
        enum:
        - percent
        - fixed
        type: string
      discount_value:
        type: integer
      expires_at:
        description: RFC3339, mis. 2026-12-31T23:59:59Z
        type: string
      max_discount:
        minimum: 0
        type: integer
      min_spend:
        minimum: 0
        type: integer
      usage_limit:
        type: integer
    required:
    - code
    - discount_type
    - discount_value
    - expires_at
    type: object
  dto.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
...
  dto.RegisterRequest:
    properties:
      confirm_password:
        type: string
      email:
        type: string
      password:
        minLength: 6
        type: string
      roles:
        description: 'Roles: pilih satu atau lebih dari buyer/seller/driver. Default
          buyer bila kosong.'
        items:
          type: string
        type: array
    required:
    - confirm_password
    - email
    - password
    type: object
  dto.SelectRoleRequest:
    properties:
      role:
        enum:
        - admin
        - buyer
        - seller
        - driver
        type: string
    required:
    - role
    type: object
...
  dto.UpdateItemRequest:
    properties:
      quantity:
        type: integer
    required:
    - quantity
    type: object
...
  response.ErrorCode:
    enum:
    - BAD_REQUEST
    - UNAUTHORIZED
    - FORBIDDEN
    - NOT_FOUND
    - CONFLICT
    - UNPROCESSABLE_ENTITY
    - INTERNAL_SERVER_ERROR
    - VALIDATION_ERROR
    - TOKEN_EXPIRED
    - TOKEN_INVALID
    type: string
    x-enum-varnames:
    - ErrCodeBadRequest
    - ErrCodeUnauthorized
    - ErrCodeForbidden
    - ErrCodeNotFound
    - ErrCodeConflict
    - ErrCodeUnprocessable
    - ErrCodeInternalServerError
    - ErrCodeValidation
    - ErrCodeTokenExpired
    - ErrCodeTokenInvalid
  response.Pagination:
    properties:
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error: {}
      error_code:
        $ref: '#/definitions/response.ErrorCode'
      list_data: {}
      message:
        type: string
      pagination:
        $ref: '#/definitions/response.Pagination'
      status:
        type: boolean
    type: object
info:
  contact:
    name: SEAPEDIA Dev
  description: |-
    Backend API marketplace multi-role (Admin, Seller, Buyer, Driver) — COMPFEST 18.
    Versi yang tampil di UI diambil dari config.Version (di-set saat runtime).
    Login pakai email. Otorisasi mengikuti ACTIVE ROLE pada JWT (pilih via /auth/select-role).
  termsOfService: http://swagger.io/terms/
  title: SEAPEDIA API
  version: "1.0"
paths:
...
  /admin/promos:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Daftar promo (admin)
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Data promo (usage_limit diabaikan)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.GenerateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Generate promo (admin)
      tags:
      - Admin
  /admin/promos/{id}:
    get:
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Detail voucher/promo by ID (admin)
      tags:
      - Admin
...
  /admin/vouchers:
    get:
      parameters:
      - description: Halaman
        in: query
        name: page
        type: integer
      - description: Item per halaman
        in: query
        name: per_page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Daftar voucher (admin)
      tags:
      - Admin
    post:
      consumes:
      - application/json
      parameters:
      - description: Data voucher
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.GenerateRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Generate voucher (admin)
      tags:
      - Admin
  /admin/vouchers/{id}:
    get:
      parameters:
      - description: Discount ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Detail voucher/promo by ID (admin)
      tags:
      - Admin
  /auth/balance-summary:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Ringkasan saldo lintas role
      tags:
      - Auth
  /auth/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: Kredensial login
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      summary: Login (pakai email)
      tags:
      - Auth
  /auth/logout:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Logout (invalidasi token via denylist)
      tags:
      - Auth
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Profil user (roles + active role)
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      parameters:
      - description: Data registrasi (login pakai email)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      summary: Registrasi user baru
      tags:
      - Auth
  /auth/select-role:
    post:
      consumes:
      - application/json
      parameters:
      - description: Role yang dipilih
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.SelectRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Pilih role aktif (terbitkan token baru)
      tags:
      - Auth
...
  /buyer/cart:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Kosongkan cart
      tags:
      - Buyer
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Lihat cart (buyer)
      tags:
      - Buyer
  /buyer/cart/items:
    post:
      consumes:
      - application/json
      parameters:
      - description: Item
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.AddItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tambah produk ke cart (single-store)
      tags:
      - Buyer
  /buyer/cart/items/{productID}:
    delete:
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hapus item dari cart
      tags:
      - Buyer
    put:
      consumes:
      - application/json
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: string
      - description: Qty baru
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateItemRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Ubah qty item cart
      tags:
      - Buyer
...
securityDefinitions:
  BearerAuth:
    description: 'Masukkan token dengan format: Bearer <token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"